projects/batch-2022-2026/160922748011-ThreatForge-AI powered vunerability detection and Auto Remediation/backend/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic>=2.7,<3.0
pydantic-settings>=2.2,<3.0
python-multipart>=0.0.9,<0.1
requests>=2.31,<3.0
httpx>=0.24.0,<1.0
google-generativeai>=0.7.0,<1.0
psycopg2-binary>=2.9.0,<3.0
faiss-cpu
numpy
groq 
sentence-transformers